// apps/sync-service/src/openapi/order.ts
export type PostV2OrderBodyOrderComment = {
  /** @nullable */
  comment?: string | null;
} | null;

export type PostV2OrderBodyOrderItemsItem = {
  ordered_qty: number;
  material_id: number;
  /** @nullable */
  reason_id?: number | null;
  /** @nullable */
  recommended_stock?: number | null;
  /** @nullable */
  other_reason?: string | null;
};

export type PostV2OrderBody = {
  type: number;
  customer_id: number;
  vendor_id: number;
  /** @nullable */
  required_date?: string | null;
  activity_id: number;
  status: number;
  /** @nullable */
  order_comment?: PostV2OrderBodyOrderComment;
  order_items: PostV2OrderBodyOrderItemsItem[];
};